Change loose expressions into a function
Previously I used two variables to: hold an array of strings which
were the weapon choices in the game, and a variable that held a randomly
chosen weapon as the computerWeapon.
In the meat of the project they ask us to create a function that returns
a string, so I converted to using a function here.